requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.6
django-celery-beat==2.7.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
exceptiongroup==1.2.2
flake8==7.1.2
iniconfig==2.0.0
kombu==5.4.2
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.10.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
redis==5.2.1
sendgrid==6.11.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
